// frontend/src/backend.d.ts
import type { Principal } from "@icp-sdk/core/principal";
export interface Some<T> {
    __kind__: "Some";
    value: T;
}
export interface None {
    __kind__: "None";
}
export type Option<T> = Some<T> | None;
export interface CarConfiguration {
    bodyColor: string;
    model: string;
    accessories: Array<string>;
    wheelStyle: string;
    interiorColor: string;
}
export interface UserProfile {
    name: string;
}
export enum UserRole {
    admin = "admin",
    user = "user",
    guest = "guest"
}
export interface backendInterface {
    assignCallerUserRole(user: Principal, role: UserRole): Promise<void>;
    getCallerUserProfile(): Promise<UserProfile | null>;
    getCallerUserRole(): Promise<UserRole>;
    getMyConfiguration(): Promise<CarConfiguration | null>;
    getUserConfiguration(user: Principal): Promise<CarConfiguration | null>;
    getUserProfile(user: Principal): Promise<UserProfile | null>;
    isCallerAdmin(): Promise<boolean>;
    listAllConfigurations(): Promise<Array<[Principal, CarConfiguration]>>;
    saveCallerUserProfile(profile: UserProfile): Promise<void>;
    saveConfiguration(config: CarConfiguration): Promise<void>;
}